pyproject.toml
[tool.poetry]
name = "new-admin-panel-sprint-3"
package-mode = false
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[[tool.poetry.source]]
name = "internal"
url = "https://pypi.org/simple/"
priority = "primary"


[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"
redis = "^5.0.4"
psycopg2-binary = "^2.9.9"
elasticsearch = "^8.13.1"
annotated-types = "^0.6.0"
pytest = "^8.2.0"


[tool.poetry.group.dev.dependencies]
ruff = "^0.4.2"
mypy = "^1.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
